Fixed 6287 - Removed EnvironmentVariableRepresenter class and refactored other classes to use this class from api-base module
