+++ api/api-pipeline-config-v8/src/main/java/com/thoughtworks/go/apiv8/admin/shared/representers/stages/JobRepresenter.java
@@ -18,9 +18,9 @@ package com.thoughtworks.go.apiv8.admin.shared.representers.stages;
 
 import com.thoughtworks.go.api.base.OutputListWriter;
 import com.thoughtworks.go.api.base.OutputWriter;
+import com.thoughtworks.go.api.representers.EnvironmentVariableRepresenter;
 import com.thoughtworks.go.api.representers.ErrorGetter;
 import com.thoughtworks.go.api.representers.JsonReader;
-import com.thoughtworks.go.apiv8.admin.shared.representers.EnvironmentVariableRepresenter;
 import com.thoughtworks.go.apiv8.admin.shared.representers.stages.artifacts.ArtifactRepresenter;
 import com.thoughtworks.go.apiv8.admin.shared.representers.stages.tasks.TaskRepresenter;
 import com.thoughtworks.go.config.ArtifactConfigs;
