+++ api/api-pipeline-config-v8/src/test/groovy/com/thoughtworks/go/apiv8/pipelineconfig/representers/PipelineConfigRepresenterTest.groovy
@@ -16,13 +16,13 @@
 
 package com.thoughtworks.go.apiv8.pipelineconfig.representers
 
+import com.thoughtworks.go.api.representers.EnvironmentVariableRepresenter
 import com.thoughtworks.go.api.util.GsonTransformer
 import com.thoughtworks.go.apiv8.admin.pipelineconfig.representers.ParamRepresenter
 import com.thoughtworks.go.apiv8.admin.pipelineconfig.representers.PipelineConfigRepresenter
 import com.thoughtworks.go.apiv8.admin.pipelineconfig.representers.TimerRepresenter
 import com.thoughtworks.go.apiv8.admin.pipelineconfig.representers.materials.MaterialsRepresenter
 import com.thoughtworks.go.apiv8.admin.pipelineconfig.representers.trackingtool.TrackingToolRepresenter
-import com.thoughtworks.go.apiv8.admin.shared.representers.EnvironmentVariableRepresenter
 import com.thoughtworks.go.apiv8.admin.shared.representers.stages.ConfigHelperOptions
 import com.thoughtworks.go.apiv8.admin.shared.representers.stages.StageRepresenter
 import com.thoughtworks.go.config.*
